:recycle: #14 feature folder approachを使った形にリファクタリング
