+++ src/stories/App.stories.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import App from '../App';
+import App from '../app/App';
 
 export default {
   title: 'App',
